Added entity data type inferencing to feature_store.py and organized code
